refactor: split out text and select input types
