+++ src/app/course/course-search-helper/course-search-helper.component.ts
@@ -1,9 +1,5 @@
-import { ENTER, SPACE } from "@angular/cdk/keycodes";
-import { Component, Input, OnInit, ViewChild } from "@angular/core";
-import { FormControl } from "@angular/forms";
-import { MatAutocompleteSelectedEvent } from "@angular/material/autocomplete";
-import { MatChipGrid, MatChipInputEvent } from "@angular/material/chips";
-import { interval, Subscription, take } from "rxjs";
+import { Component, Input, OnInit } from "@angular/core";
+import { Subject } from "rxjs";
 
 import { Displayable } from "../../_types/displayable";
 import { CourseQueryService } from "../_query/course-query.service";
@@ -29,8 +25,6 @@ export class CourseSearchHelperComponent implements OnInit {
   provider?: QueryDataProvider;
   /** All selectable compare methods, will be undefined or null when no provider */
   methods?: Displayable[];
-  /** All selectable options, will be empty array when text */
-  options: Displayable[] = [];
 
   constructor(private courseQueryService: CourseQueryService) {}
 
@@ -40,33 +34,6 @@ export class CourseSearchHelperComponent implements OnInit {
       .map(provider => ({ key: provider.key, label: provider.label }));
     this.query = this.courseQueryService.getQuery(this.index);
     if (this.query.key) this.setProvider(this.query.key, false);
-
-    // listen to value change to filter selectable options
-    this.inputControl.valueChanges.subscribe(() =>
-      this.inputValueChangeHandler(),
-    );
-  }
-
-  /** Subscription of user idle detector */
-  whenIdleSubscription?: Subscription;
-
-  /**
-   * Handler when input value changed, will trigger option fetch when user stop typing for 500ms
-   */
-  inputValueChangeHandler(): void {
-    this.setInputError(null);
-    if (this.provider?.type === "text") return;
-
-    this.whenIdleSubscription?.unsubscribe();
-    this.whenIdleSubscription = interval(500)
-      .pipe(take(1))
-      .subscribe(() => {
-        this.whenIdleSubscription?.unsubscribe();
-        this.whenIdleSubscription = undefined;
-        this.optionsPage = 0;
-        this.options = [];
-        this.getOptions(this.inputControl.value ?? "");
-      });
   }
 
   /**
@@ -83,8 +50,7 @@ export class CourseSearchHelperComponent implements OnInit {
     }
     this.provider = this.courseQueryService.getProvider(providerKey);
     this.methods = this.provider?.getMethods();
-    this.options = [];
-    this.optionsPage = 0;
+    this.providerChange.next();
   }
 
   /**
@@ -94,136 +60,22 @@ export class CourseSearchHelperComponent implements OnInit {
     this.courseQueryService.setQuery(this.index, this.query);
   }
 
+  /** notify child components that provider has changed */
+  providerChange: Subject<void> = new Subject<void>();
+
+  /**
+   * Update value after child component emit value updated event
+   * @param value emitted value from input sub component
+   */
+  valueUpdated(value: Displayable[]) {
+    this.query.value = value;
+    this.notifyQueryUpdate();
+  }
+
   /**
    * Remove this whole query from queries list
    */
   removeQuery() {
     this.courseQueryService.removeQuery(this.index);
-  }
-
-  /** Key codes that is be used to separate words in text input chips mode */
-  separatorKeyCodes = [SPACE, ENTER];
-  /** Input of value */
-  inputControl = new FormControl("");
-
-  @ViewChild("chipGrid") chipGrid!: MatChipGrid;
-  error: string | null = null;
-
-  /**
-   * Set input error to display or clear error
-   * @param error error message to set, if null, will clear error
-   */
-  setInputError(error: string | null) {
-    this.chipGrid.errorState = !!error;
-    this.error = error;
-  }
-
-  /**
-   * Add text in input to query value and clear input
-   * @param event event of input sent by MatChipInput
-   */
-  addText(event: MatChipInputEvent): void {
-    if (this.provider?.type !== "text") return;
-
-    const value = event.value.trim();
-    if (value === "") return;
-
-    const error = this.provider.getValidationResult(value);
-    this.setInputError(error);
-    if (error) return;
-
-    if (!this.query.value) this.query.value = [];
-    this.query.value.push({
-      key: event.value.trim(),
-      label: event.value.trim(),
-    });
-    this.inputControl.setValue("");
-    this.notifyQueryUpdate();
-  }
-
-  /** Is waiting for server respond options or not */
-  isLoadingOptions = false;
-  /** Visibility detector, used to additional loading options */
-  intersectionObserver = new IntersectionObserver(
-    entries => {
-      // observer will be triggered when target is going to visible or invisible
-      // so we have to ignore invisible event
-      if (!entries[0].isIntersecting) return;
-      if (this.optionsPage < 0) return;
-
-      this.isLoadingOptions = true;
-      // Stop observing while loading options
-      this.intersectionObserver.disconnect();
-      this.getOptions(this.inputControl.value ?? "");
-    },
-    { threshold: 1 },
-  );
-
-  /**
-   * Bind intersection observer to element that is preserved in options menu
-   */
-  bindLoadingTrigger(): void {
-    const target = document.querySelector(
-      ".mat-mdc-autocomplete-panel .loading-trigger",
-    );
-    if (!target) return;
-    this.intersectionObserver.disconnect();
-    this.intersectionObserver.unobserve(target);
-    this.intersectionObserver.observe(target);
-  }
-
-  /** Current page of options, needs to be reset when searching context changed */
-  optionsPage = 0;
-  /** Last query request, used to cancel last request when new request is triggered */
-  lastQueryRequest?: Subscription;
-
-  /**
-   * Get options from server
-   * @param value value to search options
-   */
-  getOptions(value: string): void {
-    this.lastQueryRequest?.unsubscribe();
-    this.lastQueryRequest = this.provider
-      ?.getOptions({
-        page: this.optionsPage + 1,
-        keyword: value.trim(),
-      })
-      .subscribe(options => {
-        this.options.push(
-          ...options.content.filter(option => {
-            return !this.query.value?.find(
-              selected => selected.key === option.key,
-            );
-          }),
-        );
-        this.optionsPage = options.pagination.current;
-        this.isLoadingOptions = false;
-        if (options.pagination.current >= options.pagination.total) {
-          this.optionsPage = -1;
-          return;
-        }
-        this.bindLoadingTrigger();
-      });
-  }
-
-  /**
-   * Select an option into query value and remove it from options list
-   * @param event event of option selected
-   */
-  selectOption(event: MatAutocompleteSelectedEvent): void {
-    if (!this.query.value) this.query.value = [];
-    this.query.value.push(event.option.value);
-    this.options.splice(this.options.indexOf(event.option.value), 1);
-    this.inputControl.setValue("");
-    this.notifyQueryUpdate();
-  }
-
-  /**
-   * Remove selected option from query value
-   * @param index index of selected option to remove
-   */
-  removeSelectedOption(index: number): void {
-    this.query.value?.splice(index, 1);
-    this.notifyQueryUpdate();
   }
 }
