+++ src/app/course/course-search-helper/select-input/select-input.component.css
@@ -1,0 +1,11 @@
+mat-form-field {
+  @apply block !important;
+}
+
+mat-chip-row {
+  @apply mx-1 !important;
+}
+
+.loading-trigger {
+  min-height: 0;
+}
