+++ src/app/course/course-search-helper/text-input/text-input.component.css
@@ -1,0 +1,3 @@
+mat-form-field {
+  display: block !important;
+}
