+++ src/app/course/course-search-helper/text-input/text-input.component.html
@@ -1,0 +1,33 @@
+<mat-form-field>
+  <mat-label>比對內容</mat-label>
+
+  <mat-chip-grid
+    #chipGrid
+    aria-label="Content Selection"
+    [disabled]="!provider"
+  >
+    <ng-container *ngFor="let selected of value; let index = index">
+      <span class="inline-block self-center" *ngIf="index > 0">或</span>
+      <mat-chip-row (removed)="removeSelectedOption(index)">
+        {{ selected.label }}
+        <button matChipRemove [attr.aria-label]="'remove ' + selected">
+          <mat-icon>cancel</mat-icon>
+        </button>
+      </mat-chip-row>
+    </ng-container>
+  </mat-chip-grid>
+
+  <input
+    matInput
+    type="text"
+    placeholder="輸入來過濾或選擇"
+    (focus)="active.emit()"
+    [matChipInputFor]="chipGrid"
+    matChipInputAddOnBlur
+    (matChipInputTokenEnd)="addText($event)"
+    [matChipInputSeparatorKeyCodes]="separatorKeyCodes"
+    [formControl]="inputControl"
+    [value]="inputControl.value"
+  />
+  <mat-error>{{ error }}</mat-error>
+</mat-form-field>
